// src/types/OrderInterface.ts
export interface OrderInterface {
	id: string;
	customer: string;
	product: string[];
	status: string;
	date: Date;
	// shippingDate: string;
	// shippingName: string;
	// shippingAddress: string;
	// shippingCity: string;
	// shippingZipCode: string;
	// shippingCountry: string;
	// billingName: string;
	// billingAddress: string;
	// billingCity: string;
	// billingZipCode: string;
	// billingCountry: string;
}